% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zebrafish-package.R
\name{zebrafish-package}
\alias{zebrafish-package}
\alias{zebrafish}
\title{Zebrafish: plot and data processing\if{html}{\out{
<div style="float: right; padding-left: 10px;">
  <img src="figures/logo.png" width="200"/>
</div>
}}}
\description{
This package provides tools for dealing zebrafish data and visualization.
\itemize{
\item \code{zebrafish_be}: Analyzing behavioral tests for distance and speed.
\item \code{zebrafish_plot}: Visualization of behavioral test results.
\item \code{zebrafish_enrich}: Functional enrichment analysis of zebrafish genes.
}
}
\section{安装}{

You can install it using the following command:

\if{html}{\out{<div class="sourceCode">}}\preformatted{devtools::install_github("preasepe-zyt/zebrafish-package")
}\if{html}{\out{</div>}}
}

\author{
\if{html}{\out{<b>}}zyt-preaspe\if{html}{\out{</b>}}
}
