% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ze.R
\name{zebrafish_be}
\alias{zebrafish_be}
\title{zebrafish_be.\if{html}{\out{
<div style="float: right; padding-left: 10px;">
  <img src="figures/be.png" width="200"/>
</div>
}}}
\usage{
zebrafish_be(dat, ti, ng, way)
}
\arguments{
\item{dat, }{A data frame to record behavioral data.}

\item{ti, }{Total time of behavioral test.}

\item{ng, }{Number of zebrafish group in behavioral test.}

\item{way, }{Speed or distance to analysis.}
}
\value{
A matrix.
}
\description{
Dealing with behavioral data.
}
\details{
Need to confirm dat, ti, ng, way.
}
\examples{
data(data)
results1_dis <- zebrafish_be(data[[1]], 20, 8, "dis")
results1_speed <- zebrafish_be(data[[1]], 20, 8, "speed")
results2_dis <- zebrafish_be(data[[2]], 20, 8, "dis")
results2_speed  <- zebrafish_be(data[[2]], 20, 8, "speed")
results3_dis <- zebrafish_be(data[[3]], 20, 8, "dis")
results3_speed  <- zebrafish_be(data[[3]], 20, 8, "speed")

}
