% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zebrafish_enrich.R
\name{zebrafish_enrich}
\alias{zebrafish_enrich}
\title{zebrafish_enrich.\if{html}{\out{
<div style="float: right; padding-left: 10px;">
  <img src="figures/enrich.jpg" width="200"/>
</div>
}}}
\usage{
zebrafish_enrich(dat, class)
}
\arguments{
\item{dat, }{A vector for zebrafish genes.}

\item{class, }{The type of enrichment analysis for go or kegg.}
}
\value{
Plot.
}
\description{
Conduct GO and KEGG enrichment analysis for zebrafish genes
}
\details{
Need to confirm dat, group, way.
}
\examples{
genes <- c("apoea", "appb", "psen1", "psen2", "mapta", "bin1", "clu", "picalma","trem2a", "abca7b", "cd33", "cr1l", "sorl1", "grn1", "sod1", "hspb1","gsk3ba", "bace1", "tnfa", "il6", "nlrp3", "cx3cr1")
zebrafish_enrich(genes,"go")
zebrafish_enrich(genes,"kegg")

}
