% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zebrafish_plot.R
\name{zebrafish_plot}
\alias{zebrafish_plot}
\title{zebrafish_plot.\if{html}{\out{
<div style="float: right; padding-left: 10px;">
  <img src="figures/plot.jpg" width="200"/>
</div>
}}}
\usage{
zebrafish_plot(dat, group, way)
}
\arguments{
\item{dat, }{A data from zebrafish_be.}

\item{group, }{A vector for zebrafish group.}

\item{way, }{Speed or distance to analysis.}
}
\value{
Plot.
}
\description{
visualization behavioral data.
}
\details{
Need to confirm dat, group, way.
}
\examples{
a <- c("Ctl", "M", "Y", "100 µg/mL", "140 µg/mL", "180 µg/mL")
zebrafish_plot(results1_dis,a,"dis")
zebrafish_plot(results1_speed,a,"speed")

}
